+++ src/ui/View.h
@@ -33,7 +33,7 @@ public:
     void timerCallback() override;
 
     void paint(Graphics& g) override;
-    void drawWave(Graphics& g, std::vector<double>& samples, Colour color) const;
+    void drawWave(Graphics& g, std::vector<float>& samples, Colour color) const;
     void drawGrid(Graphics& g);
     void drawSegments(Graphics& g, bool isMainPattern);
     void drawMidPoints(Graphics& g);
